requirements.txt
click==8.1.8
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pillow==11.2.1
Pygments==2.19.1
pytesseract==0.3.13
python-magic==0.4.27
PyYAML==6.0.2
rich==14.0.0
shellingham==1.5.4
typer==0.15.4
typing_extensions==4.13.2
watchdog==6.0.0
